Move bool helper into shared helpers module
